Modify covariance matrix convergence calculation
    modified:   api-demo_centroid.csv
    modified:   api-demo_predict.csv
    modified:   glassfire_centroid.h
    modified:   test_classifier.cpp

+++ test_classifier.cpp
@@ -14,12 +14,12 @@ using namespace std;
 int main(){
 
 
-    typedef glassfire::Classifier<float,2, size_t> Classifier;
+    typedef glassfire::Classifier<double,2, size_t> Classifier;
     auto feature_s  = Classifier::Feature_s();
     auto classifier = Classifier();
 
     size_t Index;
-    float dummy, axis_0, axis_1;
+    double dummy, axis_0, axis_1;
 
     cout << "-- Load feature from file ..." << endl;
     auto infile = ifstream("rand.csv");
